refactor(PotreeDebug): rename PotreeDebug to PointCloudDebug
