+++ examples/copc_simple_loader.html
@@ -109,7 +109,7 @@
                     pointBudget: 3000000,
                 });
                 view.addLayer(layer).then(onLayerReady);
-                debug.PotreeDebug.initTools(view, layer, gui);
+                debug.PointCloudDebug.initTools(view, layer, gui);
             }
 
             function loadAutzen() {
