+++ examples/entwine_3d_loader.html
@@ -98,7 +98,7 @@
 
                 itowns.View.prototype.addLayer.call(view, eptLayer).then(onLayerReady);
 
-                debug.PotreeDebug.initTools(view, eptLayer, debugGui);
+                debug.PointCloudDebug.initTools(view, eptLayer, debugGui);
             }
 
             readEPTURL();
