+++ examples/potree_25d_map.html
@@ -102,7 +102,7 @@
                     potreeLayer.root.bbox.getSize(size);
                     potreeLayer.root.bbox.getCenter(lookAt);
 
-                    debug.PotreeDebug.initTools(view, potreeLayer, debugGui);
+                    debug.PointCloudDebug.initTools(view, potreeLayer, debugGui);
 
                     view.camera3D.far = 2.0 * size.length();
