+++ examples/potree_3d_map.html
@@ -71,7 +71,7 @@
 
             // add potreeLayer to scene
             function onLayerReady() {
-                debug.PotreeDebug.initTools(view, potreeLayer, debugGui);
+                debug.PointCloudDebug.initTools(view, potreeLayer, debugGui);
 
                 // update stats window
                 var info = document.getElementById('info');
