+++ utils/debug/Main.js
@@ -1,5 +1,5 @@
 export { default as Debug } from './Debug';
-export { default as PotreeDebug } from './PotreeDebug';
+export { default as PointCloudDebug } from './PointCloudDebug';
 export { default as createTileDebugUI } from './TileDebug';
 export { default as create3dTilesDebugUI } from './3dTilesDebug';
 export { default as GeometryDebug } from './GeometryDebug';
